process_pub_database/requirements.txt
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
idna==2.8
lxml==4.4.0
pymediawiki==0.6.1
requests==2.22.0
soupsieve==1.9.2
urllib3==1.25.3
